Preparation for full user login support
Store nickname of logged in user in session. Define allowEntry earlier so that including includes/loginFunctions.php will not cause a 403. Fixed includes/loginFunctions.php#loadCookie's broken query and method call to a null object. Use isset instead of array_key_exists. Give HTTP 403 with improper GET/POST keys in src/register.php#commit. Give HTTP 403 when opening src/register.php while already logged in. Changed id of Register button in src/register.php and renamed src/register.css to src/formpage.css so that its rules will be able to be used on the login page too. Added a checkbox for persistent login and a link for password reset to login form on header. Added maxlengths to fields on login form in header. Changed the form id of the header's login form so that its CSS rules could be used for the user profile box in the header when a user is logged in.