Implemented login submission and login form
Pressing the Login button on the header or on src/login.php will now do
the backend stuff for cookies and sessions. src/login.php without any
POST variables will display a full page login form for when the user
mistypes their email address or password. Header's input#persistent id
changed to input#remember to prevent conflict with input#persistent in
src/login.php. Swap out the login box for a user profile box in the
header when a user is associated with the session.