Titles and authors on posts
Added `title` field to `posts` table and changed type of `posttime`
field for easier manipulation in PHP. Display author and title on posts.
Changed register button to register link and added missing table column
on header form to fix HTML5 validation. Fixed PHP errors showing when
attempting to log in with an unregistered email address. Prevent users
from registering with the nickname "Anonymous" so that posts without an
author cannot be confused for posts by a user with that name. Explicitly
assign style properties to several elements in notes in preparation for
compose system that will have a different set of element types that
share the same styles as the existing ones.