Implemented post compose
Show a compose sticky note that is styled after other notes with the
exception of color palette. index.php now will commit a note when
newtitle and newnote POST variables are present, and will either print
out only the newly created note (for AJAX) or the entire updated page
(for reloading the page after a POST) depending on whether the client
has JavaScript enabled or not. If the client has JavaScript enabled, the
compose note and its contents will autosize when modified and
subsequently rearrange surroundings notes when focus is lost so that
there are absolutely no scrollbars; otherwise, compose note stays a
constant size and will have scrollbars in its message body area if
enough lines are typed.