Backup cookie token generator; whitespace changes
Use Mersenne Twister as a less secure but ubiquitous fallback for cookie
token generation if the php_openssl extension is not available.
Re-enable pre-wrap white-space and do not trim anything entered into the
compose note so that exact spacing that was entered into the compose
note is preserved when displayed as a regular note. Maximize amount of
characters per note by storing notes unescaped in database and then
escaping HTML and replacing newlines with line breaks in
index.php#constructNoteElement. Grow board when compose note autosizes
beyond board's height.